Extract common date-fns adapter logic into a base adapter
