+++ packages/x-date-pickers/src/AdapterDateFnsBase/index.ts
@@ -1,0 +1,1 @@
+export { AdapterDateFnsBase } from './AdapterDateFnsBase';
