+++ packages/x-date-pickers/src/internals/models/helpers.ts
@@ -10,6 +10,8 @@ export type ExtendMui<C, Removals extends keyof C = never> = Omit<
 
 export type MakeOptional<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
 
+export type MakeRequired<T, K extends keyof T> = T & Required<Pick<T, K>>;
+
 export type DefaultizedProps<
   P extends {},
   RequiredProps extends keyof P,
